Update patch set 1
Patch Set 1: Code-Review+1

The LG v400 uses the same Open Bump script in CM sources and it would be much better if the v500 uses it also.  Loki has an ugly error message upon startup and requires a backported aboot.img.  The bumped kernels support the 4.4.2 aboot.img and allow us to upgrade our firmware and bootloader to a newer version.

Patch-set: 1