Update patch set 1
Patch Set 1:

Loki still works with 4.2.2 bootloader, but not with 4.4.2 bootloader. Although, I don't see any problems with transitioning from Loki to a bumped kernel. The bumped kernel can be installed over the top of the Loki kernel without problems using the extract command. There is currently no auto Loki happening during the installation of cm. Currently, the Rom installation updater-script calls the Loki shell script to lokify the kernel upon installation. All we do is remove that code from the installet-script that calls the Loki script (which the patch above does).

Patch-set: 1