Update patch set 1
Patch Set 1:

But what I'm saying is that the recovery has never had to decide if it needs to modify the boot.img during the installation of cm. The installer-script that comes packaged with the actual Rom tells the recovery if the boot image needs Loki or not. When we modify the install script to remove the Loki call, the recovery will know that the boot image is already bumped. The new cyanogen recovery doesn't even have an auto Loki option.

Patch-set: 1