Update patch set 1
Patch Set 1:

Ahh, I was not aware of the previous deprecated method. I've always wondered what auto Loki did. :-) Thanks for your feedback. It would be really nice if this was merged. I don't really like the secure boot error message on startup very much and it would be really nice to not have to backport aboot.img from 4.2.2. Once Loki is removed, we can have a pure 4.4.2 bootloader and aboot.

Patch-set: 1