Update patch set 1
Patch Set 1: -Code-Review

Ethan, the problem with the Bluetooth service is that it gets a new mac address every time the service is restarted.  Any idea how to prevent that? I have been using audit2allow to generate the rules, but I'm not exactly sure about the rules for creating new labels.

Patch-set: 1