Update patch set 1
Patch Set 1:

By default, selinux blocks everything.  The purpose of the sepolicy is only to allow actions, not deny actions.  By removing some of the sepolicy rules, it is only making the sepolicy more enforcing.

Also, the service that sets the bluetooth address is running as "root" user, which is how almost every device does it.

Patch-set: 1