Update patch set 1
Patch Set 1: -Code-Review

This patch is confirmed to fix the problem with the bt address changing when the service is restarted. So, at least we now know that this is the last rule that needs to be fixed for Bluetooth. :-)

How can I determine which system_prop and which block_device is being used?

Patch-set: 1