Update patch set 6
Patch Set 6:

Ethan,

I checked the hex of libqc-opt and it is aosp powerhal, no doubt about it. The v510 is built using mako sources with very little modification.

When comparing the aosp and qcom powerhal code, they have different paths for stuff.

When I look at the hex from libqc-opt, the paths are from aosp code.

The size of the powerhal is the same as in aosp.

In v500 with 30a firmware, the size of powerhal is the same as qcom and the code you see in hex is the same.

qcom powerhal goes to /system/vendor/lib/hw
aosp powerhal goes to /system/lib/hw

That should be good evidence it is AOSP HAL with the v510 libqc-opt we are using now.

Patch-set: 6