Update patch set 1
Patch Set 1: -Code-Review

I think the main idea is to go back to the default configuration as much as possible.  Changing gCrdaDefaultCountryCode back to 00 might be ok.  But, there are also reports that using gCrdaDefaultCountryCode=00 still does not allow connecting to some 5GHg networks.  The best solution would be to be able to manually select a specific country code in the network settings.

Patch-set: 1