Update patch set 1
Patch Set 1:

> During my testing, I've noticed that file transfers slow down
 > significantly when the device goes to sleep mode.

That may be because of automatic tx power control. Have you tried disabling it?


 > Yes, I think the ini file changes need more scrutiny in general.
 > After all, both the changes you mentioned were shot down before.
 > 
 > https://review.cyanogenmod.org/#/c/175812/
 > https://review.cyanogenmod.org/#/c/175813/

Yes, but that was with the old driver and old wcnss bin file. The changes in the ini file come from caf, and also match the values flo uses. If you check other devices, most of them enable automatic tx power control and filter Mcast and Bcast.

Patch-set: 1